Made service nodes callable as pure python objects, skipping the need to call them as services
