+++ behaviours/smarc_bt/smarc_bt/mission/ros_waypoint.py
@@ -3,15 +3,15 @@
 from .waypoint import UnderwaterWaypoint
 from smarc_mission_msgs.msg import GotoWaypoint
 
-class SMaRCWP(UnderwaterWaypoint):
+class ROSWP(UnderwaterWaypoint):
     def __init__(self, goto_waypoint: GotoWaypoint) -> None:
         super().__init__()
 
-        self._goto_wp = goto_waypoint
+        self.goto_wp = goto_waypoint
 
     @property
     def name(self) -> str:
-        return self._goto_wp.name
+        return self.goto_wp.name
 
     @property
     def is_actionable(self) -> bool:
@@ -26,24 +26,24 @@ class SMaRCWP(UnderwaterWaypoint):
     
     @property
     def position(self) -> tuple[float, float, float]:
-        p = self._goto_wp.pose.pose.position
+        p = self.goto_wp.pose.pose.position
         return (p.x, p.y, p.z)
     
     @property
     def reference_frame(self) -> str:
-        return self._goto_wp.pose.header.frame_id
+        return self.goto_wp.pose.header.frame_id
     
     @property
     def travel_depth(self) -> float:
-        return self._goto_wp.travel_depth
+        return self.goto_wp.travel_depth
     
     @property
     def travel_altitude(self) -> float:
-        return self._goto_wp.travel_altitude
+        return self.goto_wp.travel_altitude
     
     @property
     def arrival_heading(self) -> float:
-        if self._goto_wp.use_heading:
-            return self._goto_wp.arrival_heading
+        if self.goto_wp.use_heading:
+            return self.goto_wp.arrival_heading
         else: 
             return None
